# use zerolog.LevelWriter interface for performance