--- a/writer_test.go
+++ b/writer_test.go
@@ -1,6 +1,8 @@
 package zlogsentry
 
 import (
+	"errors"
+	"io"
 	"testing"
 	"time"
 
@@ -38,6 +40,125 @@ func TestParseLogEvent(t *testing.T) {
 	assert.Equal(t, "bee07485-2485-4f64-99e1-d10165884ca7", ev.Extra["requestId"])
 }
 
+func TestParseLogLevel(t *testing.T) {
+	w, err := New("")
+	require.Nil(t, err)
+
+	level, err := w.parseLogLevel(logEventJSON)
+	require.Nil(t, err)
+	assert.Equal(t, zerolog.ErrorLevel, level)
+}
+
+func TestWrite(t *testing.T) {
+	beforeSendCalled := false
+	writer, err := New("", WithBeforeSend(func(event *sentry.Event, hint *sentry.EventHint) *sentry.Event {
+		assert.Equal(t, sentry.LevelError, event.Level)
+		assert.Equal(t, "test message", event.Message)
+		require.Len(t, event.Exception, 1)
+		assert.Equal(t, "dial timeout", event.Exception[0].Value)
+		assert.True(t, time.Since(event.Timestamp).Minutes() < 1)
+		assert.Equal(t, "bee07485-2485-4f64-99e1-d10165884ca7", event.Extra["requestId"])
+		beforeSendCalled = true
+		return event
+	}))
+	require.Nil(t, err)
+
+	var zerologError error
+	zerolog.ErrorHandler = func(err error) {
+		zerologError = err
+	}
+
+	// use io.MultiWriter to enforce using the Write() method
+	log := zerolog.New(io.MultiWriter(writer)).With().Timestamp().
+		Str("requestId", "bee07485-2485-4f64-99e1-d10165884ca7").
+		Logger()
+	log.Err(errors.New("dial timeout")).
+		Msg("test message")
+
+	require.Nil(t, zerologError)
+	require.True(t, beforeSendCalled)
+}
+
+func TestWriteLevel(t *testing.T) {
+	beforeSendCalled := false
+	writer, err := New("", WithBeforeSend(func(event *sentry.Event, hint *sentry.EventHint) *sentry.Event {
+		assert.Equal(t, sentry.LevelError, event.Level)
+		assert.Equal(t, "test message", event.Message)
+		require.Len(t, event.Exception, 1)
+		assert.Equal(t, "dial timeout", event.Exception[0].Value)
+		assert.True(t, time.Since(event.Timestamp).Minutes() < 1)
+		assert.Equal(t, "bee07485-2485-4f64-99e1-d10165884ca7", event.Extra["requestId"])
+		beforeSendCalled = true
+		return event
+	}))
+	require.Nil(t, err)
+
+	var zerologError error
+	zerolog.ErrorHandler = func(err error) {
+		zerologError = err
+	}
+
+	log := zerolog.New(writer).With().Timestamp().
+		Str("requestId", "bee07485-2485-4f64-99e1-d10165884ca7").
+		Logger()
+	log.Err(errors.New("dial timeout")).
+		Msg("test message")
+
+	require.Nil(t, zerologError)
+	require.True(t, beforeSendCalled)
+}
+
+func TestWrite_Disabled(t *testing.T) {
+	beforeSendCalled := false
+	writer, err := New("",
+		WithLevels(zerolog.FatalLevel),
+		WithBeforeSend(func(event *sentry.Event, hint *sentry.EventHint) *sentry.Event {
+			beforeSendCalled = true
+			return event
+		}))
+	require.Nil(t, err)
+
+	var zerologError error
+	zerolog.ErrorHandler = func(err error) {
+		zerologError = err
+	}
+
+	// use io.MultiWriter to enforce using the Write() method
+	log := zerolog.New(io.MultiWriter(writer)).With().Timestamp().
+		Str("requestId", "bee07485-2485-4f64-99e1-d10165884ca7").
+		Logger()
+	log.Err(errors.New("dial timeout")).
+		Msg("test message")
+
+	require.Nil(t, zerologError)
+	require.False(t, beforeSendCalled)
+}
+
+func TestWriteLevel_Disabled(t *testing.T) {
+	beforeSendCalled := false
+	writer, err := New("",
+		WithLevels(zerolog.FatalLevel),
+		WithBeforeSend(func(event *sentry.Event, hint *sentry.EventHint) *sentry.Event {
+			beforeSendCalled = true
+			return event
+		}))
+	require.Nil(t, err)
+
+	var zerologError error
+	zerolog.ErrorHandler = func(err error) {
+		zerologError = err
+	}
+
+	log := zerolog.New(writer).With().Timestamp().
+		Str("requestId", "bee07485-2485-4f64-99e1-d10165884ca7").
+		Logger()
+	log.Err(errors.New("dial timeout")).
+		Msg("test message")
+
+	require.Nil(t, zerologError)
+	require.False(t, beforeSendCalled)
+}
+
 func BenchmarkParseLogEvent(b *testing.B) {
 	w, err := New("")
 	if err != nil {
@@ -49,7 +170,7 @@ func BenchmarkParseLogEvent(b *testing.B) {
 	}
 }
 
-func BenchmarkParseLogEvent_DisabledLevel(b *testing.B) {
+func BenchmarkParseLogEvent_Disabled(b *testing.B) {
 	w, err := New("", WithLevels(zerolog.FatalLevel))
 	if err != nil {
 		b.Errorf("failed to create writer: %v", err)
@@ -93,7 +214,7 @@ func BenchmarkWriteLogLevelEvent(b *testing.B) {
 	}
 }
 
-func BenchmarkWriteLogLevelEvent_DisabledLevel(b *testing.B) {
+func BenchmarkWriteLogLevelEvent_Disabled(b *testing.B) {
 	w, err := New("", WithLevels(zerolog.FatalLevel))
 	if err != nil {
 		b.Errorf("failed to create writer: %v", err)