If we log something (e.g. as local log level got set to debug) but we only want to send events for Error to sentry, each event is parsed to check for the log level.

As there is the LevelWriter interface as zerolog suggests for multiple outputs [here](https://github.com/rs/zerolog#multiple-log-output). That way we might skip parsing the event if the log level is disabled.

This PR implements this.

For reference I implemented benchmarks as well. This are the results on my device
```
# make benchmarks
[...]
pkg: github.com/archdx/zerolog-sentry
BenchmarkParseLogEvent-8                          272346              3806 ns/op            5136 B/op         14 allocs/op
BenchmarkParseLogEvent_DisabledLevel-8            253910              4014 ns/op            5136 B/op         14 allocs/op
BenchmarkWriteLogEvent-8                           65269             21795 ns/op           12586 B/op        109 allocs/op
BenchmarkWriteLogEvent_Disabled-8               20960934             57.01 ns/op               0 B/op          0 allocs/op
BenchmarkWriteLogLevelEvent-8                      68061             16895 ns/op           12618 B/op        109 allocs/op
BenchmarkWriteLogLevelEvent_DisabledLevel-8    167343889             6.876 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/archdx/zerolog-sentry        9.532s
```